# Keep larger merge buffers for RPC